CBMEM: Unify get_top_of_ram()
Reviewed-on: http://review.coreboot.org/3904
Tested-by: build bot (Jenkins)

+++ src/northbridge/intel/gm45/ram_calc.c
@@ -26,6 +26,7 @@
 #include <arch/io.h>
 #include <device/pci_def.h>
 #include <console/console.h>
+#include <cbmem.h>
 #include "gm45.h"
 
 /** Decodes used Graphics Mode Select (GMS) to kilobytes. */
@@ -83,7 +84,7 @@ u32 decode_igd_gtt_size(const u32 gsm)
 	}
 }
 
-u32 get_top_of_ram(void)
+unsigned long get_top_of_ram(void)
 {
 	const pci_devfn_t dev = PCI_DEV(0, 0, 0);
